requirements.txt
tensorflow-gpu==1.6
numpy
scipy
opencv-python
tqdm
pyyaml
flake8
jupyter
